Move granular permissions to permissions folder
