+++ front/app/containers/Admin/projects/project/permissions/granular_permissions/components/UserFieldSelection/UserFieldSelection.tsx
