+++ front/app/containers/Admin/projects/project/permissions/granular_permissions/containers/Granular/index.tsx
