+++ front/app/modules/index.ts
@@ -1,7 +1,7 @@
 import { loadModules } from 'utils/moduleUtils';
 
 import smartGroupsConfiguration from './commercial/smart_groups';
-import granularPermissionsConfiguration from './commercial/granular_permissions';
+import granularPermissionsConfiguration from '../containers/Admin/projects/project/permissions/granular_permissions';
 import ideaAssignmentConfiguration from './commercial/idea_assignment';
 import moderationConfiguration from './commercial/moderation';
 import flagInappropriateContentConfiguration from './commercial/flag_inappropriate_content';
